// src/components/models/types.ts
export enum LightColor {
  RED = 'RED',
  YELLOW = 'YELLOW',
  GREEN = 'GREEN',
  FLASHING_ORANGE = 'FLASHING_ORANGE',
}

export enum Direction {
  NORTH = 'NORTH',
  SOUTH = 'SOUTH',
  EAST = 'EAST',
  WEST = 'WEST',
}

export enum LaneType {
  LEFT = 'LEFT',
  STRAIGHT_INNER = 'STRAIGHT_INNER',
  STRAIGHT_OUTER = 'STRAIGHT_OUTER',
  RIGHT = 'RIGHT',
}